requirements.txt
# Web 框架
Flask~=3.0.0
gunicorn~=23.0.0
Flask-Limiter~=3.5.0
pytest~=8.3.0

# 图像处理
Pillow~=10.4.0
piexif~=1.1.3  # 仅当需要额外处理 EXIF 时
tomli~=2.2.1
